Added the ability to list templates from multiple catalog (#74)
